# Remove extra copy in deepseek forward absorb