<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Benchmark

### Performance

main branch:
```
batch size: 1
latency: 10.73 s
output throughput: 47.70 token/s
(input + output) throughput: 95.40 token/s

batch size: 32
latency: 20.00 s
output throughput: 819.37 token/s
(input + output) throughput: 1638.74 token/s
```

this PR:
```
batch size: 1
latency: 10.67 s
output throughput: 47.97 token/s
(input + output) throughput: 95.95 token/s

batch size: 32
latency: 19.69 s
output throughput: 832.17 token/s
(input + output) throughput: 1664.35 token/s
```

0.5% improvement for bs 1 and 1.5% improvement for bs 32.

main branch:
<img width="779" alt="image" src="https://github.com/user-attachments/assets/6dc4cc57-0e9c-4d34-afaa-52495f8e5dc5" />

this PR:
<img width="740" alt="image" src="https://github.com/user-attachments/assets/1018d709-0446-44c7-8ddf-fc256fb6914f" />

removed 3 copy with 2 cat added (can overlap one later), duration 40μm -> 38μm

## Accuracy
```
python3 -m sglang.launch_server --model deepseek-ai/DeepSeek-V3-0324 --tp 8 --trust-remote-code --attention-backend fa3 --disable-radix
python3 benchmark/gsm8k/bench_sglang.py --num-questions 1400 --parallel 1400 --num-shots 8
```
main barnch:
```
Accuracy: 0.947
Invalid: 0.000
Latency: 194.902 s
Output throughput: 737.433 token/s
```
this PR:
```
Accuracy: 0.949
Invalid: 0.000
Latency: 195.182 s
Output throughput: 729.388 token/s
```
this PR with deepgemm and deepgemm_bmm:
```
Accuracy: 0.944
Invalid: 0.000
Latency: 361.141 s
Output throughput: 398.598 token/s
```