@Fridge003 2025-04-20
Could you please add results for accuracy tests