@Alcanderian 2025-04-20
At first glance it seems fine, but I think compatibility with use_deep_gemm_bmm should be tested.