@ispobock 2025-04-21
@HaiShaw The AMD test seems always failed, could you help check?