@ispobock 2025-04-22
The AMD test passed, thanks @HaiShaw @saienduri !